+++ tests/EventListener/BreadcrumbListenerControllerResolutionTest.php
@@ -1,13 +1,14 @@
 <?php
 
-namespace APY\BreadcrumbTrailBundle\EventListener;
+namespace TomvdPeet\BreadcrumbBundle\Tests\EventListener;
 
 require_once __DIR__.'/../Fixtures/ControllerWithAttributes.php';
 require_once __DIR__.'/../Fixtures/InvokableControllerWithAttributes.php';
 
-use APY\BreadcrumbTrailBundle\BreadcrumbTrail\Trail;
-use APY\BreadcrumbTrailBundle\Fixtures\ControllerWithAttributes;
-use APY\BreadcrumbTrailBundle\Fixtures\InvokableControllerWithAttributes;
+use TomvdPeet\BreadcrumbBundle\BreadcrumbTrail\Trail;
+use TomvdPeet\BreadcrumbBundle\EventListener\BreadcrumbListener;
+use TomvdPeet\BreadcrumbBundle\Tests\Fixtures\ControllerWithAttributes;
+use TomvdPeet\BreadcrumbBundle\Tests\Fixtures\InvokableControllerWithAttributes;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\RequestStack;